Moved a forMSIL block.
Having large blocks of code in spots where they are never
run but where we pay their entire cost in terms of method
size is a bad idea.  If we must have these "forMSIL"
blocks, especially in hot locations like the Typer
and GenICode, the logic should all be sequestered in a
separate method to minimize the hotspot damage.