Help the parser understand more types.
It's one of those annoying inconsistencies, that the parser
won't let you write down the type of this expression in a
straightforward way.

  class X { object Y }
  (new X).Y

To add insult to injury, scala prints the type the way you
would expect to write it, but what's good for the goose is no
good for the gander it seems.

  scala> (new X).Y
  res0: X#Y.type = X$Y$@7d9e9114

The type can be expressed with existentials, so the existence
of the type isn't the issue.

  x.Y.type forSome { val x: X }

Now it can be expressed the same way scala does. We can see
that these are equivalent:

  scala> var x: x.Y.type forSome { val x: X } = _
  x: x.Y.type forSome { val x: X } = null

  scala> var x: X#Y.type = _
  x: x$1.Y.type forSome { val x$1: X } = null

Unfortunately (new X).Y cannot be assigned to either of those
due to a bug in subtyping of the well-known class where
TypeRef(pre, objectSymbol, Nil) and SingleType(pre, objectSymbol)
are handled differently.