SI-7536 track more bases when validating variance.
The variance validator was tracking only the most recently
seen definition (as "base") and calculating the relative variance
of an enclosing type parameter by traversing the owner chain from
"base" through the owner of the type parameter.

The problem with this is that it meant the variance arrow would
only be reversed at definition sites. So this would compile:

  trait Base[A]
  trait Ord[-A, A1 >: A] extends Base[A1]

because when assessing the variance of A in parent Base, the arrow
would be flipped at definition A1 for A being in the lower bound.
However this equivalent (or if anything, more restrictive) definition
would not compile:

  trait Ord[-A] extends Base[A1 forSome { type A1 >: A }]

Though A still appears in the lower bound of the type argument
to Base, the arrow reversal was not detected.