Stop throwing ValidateExceptions.
Throwing an exception is the least effective way to communicate
that one's intended invariants are not holding. For one thing it
means you will only hear about one. It's a lot easier to see what
is going on when you can hear about them all. Plus we're talking
about positions here, not a moon landing. The civilized thing to
do is issue a warning.