Fix some invalid positions emerging from Typers.
Using "setPos" without great care tends to create Trees which
violate position invariants. In this case there were Apply trees
which had among their children a Select tree with a larger range.
This was masked because at some point after the invalid position
was set but before the position validator ran, it would be
altered. But violating invariants and then cleaning up afterward
is not the way to go. Let us have invariants which do not vary.