Easier facility for testing positions.
With this you can write a partest "object Test extends PositionTest"
and then have the rest of the same file used as the source code for
which positions will be printed.

It's not ideal but it is a lot better than everyone duplicating
a bunch of DirectCompiler support code over and over, and all the
code for which positions are being tested being embedded in
Strings, which has too many issues to count.