SI-7281 1: Add @scala.runtime.ValueType to value type classes
... if it exists.

Instead of preserving AnyVal as a superclass, which seems to be out of
reach for the time being, add a marker annotation which gives runtimes
an cheap and easy way to check whether a certain class is a value type
without depending on a language-specific library.

Care is taken to handle the absence of the class (like when running on
2.10). This commit just adds the handling. The next commit will add the
class itself.